--- src/features/input/elements/Button/handleInput.copy.js
@@ -1,13 +1,0 @@
-const handleInputCopy = () => {
-    const source = document.querySelector('#input-number')
-
-    sessionStorage.setItem('sequenceNumber', source.getAttribute('value'))
-
-    const parent = document.querySelector('main')
-
-    import('../../../output/Output').then(({ default: Output }) =>
-        parent.replaceChildren(Output())
-    )
-}
-
-export default handleInputCopy
