--- src/features/output/elements/Button/handleOutput.copy.js
@@ -1,9 +1,0 @@
-const handleOutputCopy = () => {
-    const parent = document.querySelector('main')
-
-    import('../../../input/Provider').then(({ default: Provider }) =>
-        parent.replaceChildren(Provider())
-    )
-}
-
-export default handleOutputCopy
